[Refactor] #5: Common 패키지 생성 및 이동
